Applied some suggestions of splint
There are still issues with incompatible types
Cmake receipe adapted to create stricter warnings interpret as errors

+++ src/piConfigParser/piConfigParser.c
@@ -311,7 +311,7 @@ int32_t parse_modbus_master_config_data(const char* pc8_pi_config_data_p, struct
     get_json_devices_array(pc8_pi_config_data_p, &devices_array);
         
     //search for matching device types
-    for (int32_t i = 0; i < array_list_length(devices_array); i++)
+    for (size_t i = 0; i < array_list_length(devices_array); i++)
     {
         json_object *json_pi_device = NULL;		
         const char* productType = NULL;
@@ -438,7 +438,7 @@ int32_t parse_modbus_slaves_config_data(const char* pc8_pi_config_data_p, struct
     }
         
     //search for matching device types
-    for (int32_t i = 0; i < array_list_length(devices_array); i++)
+    for (size_t i = 0; i < array_list_length(devices_array); i++)
     {
         json_object *pi_device = NULL;		
         const char* productType = NULL;
